+++ sqlstore/sqlstore_test.go
@@ -30,7 +30,7 @@ var (
 	testDBPort       int
 
 	tableNames = [...]string{"ledger", "accounts", "parties", "assets", "blocks", "node_signatures",
-		"erc20_multisig_signer_events", "trades", "market_data", "orders"}
+		"erc20_multisig_signer_events", "trades", "market_data", "orders", "margin_levels", "liquidity_provisions"}
 	postgresServerTimeout = time.Second * 10
 )