+++ sqlsubscribers/margin_levels_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 	"time"
 
+	"code.vegaprotocol.io/data-node/entities"
 	"code.vegaprotocol.io/data-node/logging"
 	"code.vegaprotocol.io/data-node/sqlsubscribers"
 	"code.vegaprotocol.io/data-node/sqlsubscribers/mocks"
@@ -13,6 +14,31 @@ import (
 	"code.vegaprotocol.io/vega/types/num"
 	"github.com/golang/mock/gomock"
 )
+
+func TestMarginLevelsDuplicate_Push(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	store := mocks.NewMockMarginLevelsStore(ctrl)
+
+	store.EXPECT().Add(gomock.Any()).Times(2)
+	store.EXPECT().Flush(gomock.Any()).Times(4)
+
+	accountSource := TestAccountSource{}
+
+	subscriber := sqlsubscribers.NewMarginLevels(store, accountSource, logging.NewTestLogger())
+	subscriber.Push(context.Background(), events.NewTime(context.Background(), time.Now()))
+	subscriber.Push(context.Background(), events.NewMarginLevelsEvent(context.Background(), types.MarginLevels{}))
+	subscriber.Push(context.Background(), events.NewTime(context.Background(), time.Now()))
+	subscriber.Push(context.Background(), events.NewMarginLevelsEvent(context.Background(), types.MarginLevels{}))
+	subscriber.Push(context.Background(), events.NewTime(context.Background(), time.Now()))
+
+	// Now push a non duplicate
+
+	subscriber.Push(context.Background(), events.NewMarginLevelsEvent(context.Background(), types.MarginLevels{InitialMargin: num.NewUint(6)}))
+	subscriber.Push(context.Background(), events.NewTime(context.Background(), time.Now()))
+
+}
 
 func TestMarginLevels_Push(t *testing.T) {
 	ctrl := gomock.NewController(t)
@@ -24,7 +50,7 @@ func TestMarginLevels_Push(t *testing.T) {
 	store := mocks.NewMockMarginLevelsStore(ctrl)
 
 	store.EXPECT().Add(gomock.Any()).Times(1)
-	store.EXPECT().Flush(gomock.Any()).Times(1)
+	store.EXPECT().Flush(gomock.Any()).Times(2)
 	subscriber := sqlsubscribers.NewMarginLevels(store, accountSource, logging.NewTestLogger())
 	subscriber.Push(context.Background(), events.NewTime(context.Background(), time.Now()))
 	subscriber.Push(context.Background(), events.NewMarginLevelsEvent(context.Background(), types.MarginLevels{
@@ -37,4 +63,17 @@ func TestMarginLevels_Push(t *testing.T) {
 		Asset:                  "DEADBEEF",
 		Timestamp:              time.Now().UnixNano(),
 	}))
+
+	subscriber.Push(context.Background(), events.NewTime(context.Background(), time.Now()))
+}
+
+type TestAccountSource struct{}
+
+func (TestAccountSource) Obtain(ctx context.Context, a *entities.Account) error {
+	a.ID = 1
+	return nil
+}
+
+func (TestAccountSource) GetByID(id int64) (entities.Account, error) {
+	panic("implement me")
 }